feat: 전체 dnd 적용 작업1
