+++ src/components/Planner/Class.tsx
@@ -2,13 +2,15 @@ import { useContext, useState } from "react";
 import BackgroundModal from "../Modal/BackgroundModal";
 import AddClassModal from "../Modal/AddSemesterModal";
 import { ClassBox, deleteClass, plusClass } from "./Class.css";
+import { Draggable } from "react-beautiful-dnd";
 
 interface ClassProps {
   className?: string;
   classType?: string;
   classCredit?: string;
   active?: boolean;
-  id?: number;
+  id: string;
+  index: number;
 }
 
 const Class = ({
@@ -16,31 +18,44 @@ const Class = ({
   classType,
   classCredit,
   id,
+  index,
   active = true,
 }: ClassProps) => {
+  const isDragDisabled = id === "1";
   return (
-    <div className={ClassBox}>
-      {active ? (
-        <>
-          <h3>{className}</h3>
-          <h5>
-            {classType} {classCredit}
-          </h5>
-          <button className={deleteClass} onClick={() => console.log(1)}>
-            x
-          </button>
-        </>
-      ) : (
-        <>
-          <button
-            className={plusClass}
-            onClick={() => console.log("plusClass")}
-          >
-            +
-          </button>
-        </>
+    <Draggable draggableId={id} index={index} isDragDisabled={isDragDisabled}>
+      {(provided, snapshot) => (
+        <div
+          className={ClassBox}
+          {...provided.dragHandleProps}
+          {...provided.dragHandleProps}
+          ref={provided.innerRef}
+          // isDragging={snapshot.isDragging}
+          //isDragDisabled={isDragDisabled}
+        >
+          {active ? (
+            <>
+              <h3>{className}</h3>
+              <h5>
+                {classType} {classCredit}
+              </h5>
+              <button className={deleteClass} onClick={() => console.log(1)}>
+                x
+              </button>
+            </>
+          ) : (
+            <>
+              <button
+                className={plusClass}
+                onClick={() => console.log("plusClass")}
+              >
+                +
+              </button>
+            </>
+          )}
+        </div>
       )}
-    </div>
+    </Draggable>
   );
 };
